Don't try to publish docs on a dependabot commit (#98)
We're having the same problem here as in the main repository.
Commits caused by dependabot updates only have a read access token because of which the push to the `gh-pages` branch failes. To fix this we're doing the same, which is to ignore dependabot commits and run the workflow using a cronjob to still update the changes from dependabot in a timely manner.